--- .github/workflows/python-app.yml
@@ -1,36 +1,0 @@
-# This workflow will install Python dependencies, run tests and lint with a single version of Python
-# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
-
-name: Veronica Model Training
-
-on:
-  push:
-    branches: [ "development" ]
-  pull_request:
-    branches: [ "development" ]
-
-permissions:
-  contents: read
-
-jobs:
-  Train-Veronica:
-
-    runs-on: ubuntu-latest
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python 3.10
-      uses: actions/setup-python@v3
-      with:
-        python-version: "3.10"
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install torch numpy joblib
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
-    - name: Train And Run Veronica
-      run: |
-        # stop the build if there are Python syntax errors or undefined names
-        python3 main.py . --count --select=E9,F63,F7,F82 --show-source --statistics
-       
-
