+++ AI/loadData.py
@@ -1,0 +1,2 @@
+import torch
+from AI import Tokenizer , Veronica
